requirements.txt
numpy==1.16.4
opencv-contrib-python==4.1.1.26
pyzmq==17.1.2
